[led] Add support for binary on/off led types. Fixes JB#31507
All currently available led backends assume multiple channels and
adjustable led brightness and thus are not suitable for controlling
leds that can simply be turned on or off.

Implement new backend that allows higher level logic that assumes
rgb color control to be available to work also with binary leds.