[quirks] Add enablers for tweaking backend behavior via config files
So far the logic has been written so that it works in as many devices
as possible without need for tuning. However this easily leads to
situations where performance is compromised in every device and makes
further code changes more difficult even if/when all possibly affected
devices are available for testing.

Implement quirk system on top of mce settings that allows use of
numerical values to be used to tweak backend behavior on specific
devices only.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>