[sysfs-led] Make all sysfs control paths configurable. MER#1811
Maintaining a set of possible sysfs control paths in the code
is problematic in the sense that it always requires new sw
version to support a new device port. Additionally probing is
more likely to produce false positive matches as the set of
supposedly "known" sysfs paths grows.

Arrange it so that when a led backend is selected via
configuration file, all of the sysfs paths it potentially needs
can also be defined in the configuration file.

As the generic ini-file parsing makes it a bit difficult to see
what kind of files are actually expected, add example ini-file
for each backend.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>