[sysfs-led] Sync structure member names with sysfs paths
The names of structure members in various led backends do not
necessarily have any resemblance to sysfs control files they
relate to.

Rename all structure member variables so that they match with
the basename of default sysfs control files. This makes it a
bit easier to follow the logic of sysfs writes and also acts as
an enabler for making the paths more configurable.

Use sysfsval interface for reading and updating sysfs states.
This implicitly avoids no-change sysfs writes and allows
debugging the writes without code modifications. Not done to
those backends where side effects from potentially skipping
updates are unknown and could not be tested.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>