[sysfs-led] Do not lock-in to wrong led reset state
When led is making transition from blinking to non-blinking state or vice
versa, the intent is that the led is briefly powered off before the new
target state is programmed. This led-reset-is-needed condition is evaluated
in wrong place so that it can get locked in to wrong value when multiple
almost simultaneous pattern changes are requested from upper levels and
led reset is skipped - which can cause unwanted off sync "rainbow" blinking
at least in Jolla phone.

Check the need for reset on all transitions. And once set, do the led
reset even if the pattern requiring the reset gets replaced with something
else before getting actually activated.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>